Fix Resharper inline rename
When using VsVim in conjunction with Resharper 23.x, inline rename and similar actions are opening in visual mode instead of the previous external edit mode.

This is caused by the fact that the ITaggingProvider used to determine the presence/version of Resharper is not exported anymore by Resharper.

This commit changes the value of ResharperTaggerProviderName to another tagging provider exported by Resharper.